Added tripconfirm variants to the dev-emulator-program and dev-emulator-study
Added variants to the original dev-emulator-program/study. Adding onto the concept that program is the trip-notes and place-notes button for TimeUseSurvey and the study is the "control".

These new variants introduce the "trip-labels" button which **if filled in** (with "TripConfirmSurvey") will display a trip confirm survey button instead of the typical Multilabel format of Mode: ??? Purpose: ???