Change the data key to match the proposed version on the phone
We should also consider where we want to store the data keys.
Currently, we store the data.key along with the trip name.

```
        "TimeUseSurvey": {
          "formPath": "https://raw.githubusercontent.com/sebastianbarry/nrel-openpath-deploy-configs/surveys-info-and-surveys-data/survey-resources/data-json/time-use-survey-form-v9.json",
          "version": 9,
          "compatibleWith": 1,
          "dataKey": "manual/timeuse_survey"
        }
      },
```

That's fine but then if we want to store a new type of survey (e.g. missing
trip survey), we cannot do so without adding the new data type to the server.
Also, what data would we pull from the sever to match?

We should re-think this approach and potentially store the key in the buttons
instead

```
      "buttons": {
        "trip-notes": {
          "surveyName": "TimeUseSurvey",
          "survey_key: "manual/trip_addition_input"
          "not-filled-in-label": {
            "en": "Add Activity",
            "es": "Añadir actividad"
          },
          "filled-in-label": {
            "en": "",
            "es": ""
          }
        },
```