Init commit for main.yml + email-config.py
main.yml will authenticate AWS with our IAM role, use a dependency to get the name of the latest file pushed to configs directory, and run email-config.py.

Pushing email-config.py as it was in the last commit to my  nataliejschultz:AWS-email-config PR.

Reverting README.md to before test push.