update showsIf conditions for conditional surveys
Finally!

First, the coordinates change from NREL Building 16 (which we were using for the purpose of testing) to DFC Building 40 (which is the actual location of the Fermata chargers).

Second, instead of using the 'sensed_mode_str' (which we were also doing for the purpose of testing), we can actually use the vehicle detected by bluetooth to show surveys conditionally.

These "showsIf" expressions work with the changes in https://github.com/e-mission/e-mission-phone/pull/1145.
It will not work in Python yet; I am planning a way to do that later.

"confirmedMode" is either a vehicle detected by bluetooth, or the label chosen by the user. In this instance it is the former; it is one of the vehicles defined under "vehicle_identities" (or, for a trip where there was no vehicle detected, detected it is undefined).

Updating config version to 3 so we can receive these updates using the in-app config refresh feature!