🔧 Collapse all bike modes to one for godcgo
GoDCgo wants to use OpenPATH to evaluate a campaign to encourage people to bike
more. As part of the campaign, they want us to support "replaced mode" for bikes.
However, we currently support only one mode of interest, and have three
separate bike types. Since supporting multiple modes of interest would require
code changes, we collapse all the bike modes into one and set that as the mode
of interest.