add second location for testing DfcEvReturnTrip
Now the DfcEvReturnTrip will show if the BLE mode is E_CAR and the end location is inside either of these two boxes.

Also, since these surveys are read in order and the matched survey is the first one whose condition matches, we can treat this like an if / else if block.
I reordered DfcEvReturnTrip before DfcEvRoamingTrip so we don't have to specifically say `!pointIsWithinBounds(...)` on DfcEvRoamingTrip, making these conditions slightly less verbose.