update custom label-options to new format
As part of the work on unifying footprint calculations, the label options spec gets the following changes:
- baseMode renamed to base_mode (to be consistent with other config fields which are snake_case)
- the new 'footprint' field, which is an object with energy intensities broken down by fuel type. This will replace 'kgCo2PerKm'.
- properties are inherited from the base_mode. For example, a mode label option with base_mode of CAR will inherit the icon, color, met, and footprint of CAR. These can still be overridden if given on the mode label option.

I have updated the existing label options to adhere to this. Many of the existing fields will no longer be used or will be redundant because the base mode already provide them. I have left these fields in the JSON files for backwards-compatibility reasons, organized onto a separate, indented line. Once we are confident that users are on a new version of the app, we can remove them. Only the fields on the unindented lines will remain.

example-program-label-options.json and example-study-label-options.json are removed. Instead of these, we are going to have the docs point to the "default" label options (https://github.com/JGreenlee/e-mission-common/blob/master/src/emcommon/resources/label-options.default.json) to give deployers an example of how to set up a label options file
the example label-options were not used by anything except the dev-emulator configs. I removed the label_options field from these, so they will just use the default label options.