remove emails of people who have left their respective organizations
For cortezebikes, I removed the outdated email (I did not add a new email since the main collection period is over)
For ride2own, we have enough other representatives from We All Rise and the The Street Trust, so I did not add anyone else.
For usaid-laos, I added myself to replace the outdated NREL emails.