fix typos and string/bool issues with `admin_dashboard.token_generate`
When I was meeting with sm-ebike, I saw that their admin dashboard had the Tokens page enabled.
To my surprise, I looked at their config and saw `"token_generate": "false"`

Then I realized it should be `false` with no quotes. A string containing the word `"false"` evaluates as a truthy value and causes the Tokens page to show up.

Looking further I realized there's actually 2 issues going on here. A lot of other programs have it misspelled as toekn_generate, which is because of a typo in parse-issue-body.js

In parse-issue-body.js I fixed both the typo and the string/boolean mixup, then I manually fixed all the affected configs.