✨ define deployment-level overrides on native configs: tracking, sync (#166)
Some behaviors of the native code are configurable via the "tracking" config and "sync" config. Defaults are set in the native code (i.e. app level) and users can edit in Profile > Developer Zone, but up to now they have not been set at the deployment level.
For the fleet version of OpenPATH, we did add support for one deployment-level config option under "tracking" ("bluetooth_only" which mapped to "isFleet" in the native code) but it was handled separately from other options.

We are reworking the native code so that *any* fields defined in "tracking" or "sync" override the built-in default values when that config is constructed.
This brings the deployment config spec in line with the native code spec and should simplify the config logic in the process.